cleanup(pubsub): rename `*Stub` files to match generator
I am planning to replace these files with generated files. Renaming
them to match what the generator creates is the first step.
