+++ google/cloud/pubsub/google_cloud_cpp_pubsub.bzl
@@ -34,23 +34,23 @@ google_cloud_cpp_pubsub_hdrs = [
     "internal/extend_leases_with_retry.h",
     "internal/flow_controlled_publisher_connection.h",
     "internal/ordering_key_publisher_connection.h",
-    "internal/publisher_auth.h",
-    "internal/publisher_logging.h",
-    "internal/publisher_metadata.h",
+    "internal/publisher_auth_decorator.h",
+    "internal/publisher_logging_decorator.h",
+    "internal/publisher_metadata_decorator.h",
     "internal/publisher_round_robin.h",
     "internal/publisher_stub.h",
     "internal/publisher_stub_factory.h",
     "internal/rejects_with_ordering_key.h",
-    "internal/schema_auth.h",
-    "internal/schema_logging.h",
-    "internal/schema_metadata.h",
+    "internal/schema_auth_decorator.h",
+    "internal/schema_logging_decorator.h",
+    "internal/schema_metadata_decorator.h",
     "internal/schema_stub.h",
     "internal/sequential_batch_sink.h",
     "internal/session_shutdown_manager.h",
     "internal/streaming_subscription_batch_source.h",
-    "internal/subscriber_auth.h",
-    "internal/subscriber_logging.h",
-    "internal/subscriber_metadata.h",
+    "internal/subscriber_auth_decorator.h",
+    "internal/subscriber_logging_decorator.h",
+    "internal/subscriber_metadata_decorator.h",
     "internal/subscriber_round_robin.h",
     "internal/subscriber_stub.h",
     "internal/subscription_batch_source.h",
@@ -99,23 +99,23 @@ google_cloud_cpp_pubsub_srcs = [
     "internal/extend_leases_with_retry.cc",
     "internal/flow_controlled_publisher_connection.cc",
     "internal/ordering_key_publisher_connection.cc",
-    "internal/publisher_auth.cc",
-    "internal/publisher_logging.cc",
-    "internal/publisher_metadata.cc",
+    "internal/publisher_auth_decorator.cc",
+    "internal/publisher_logging_decorator.cc",
+    "internal/publisher_metadata_decorator.cc",
     "internal/publisher_round_robin.cc",
     "internal/publisher_stub.cc",
     "internal/publisher_stub_factory.cc",
     "internal/rejects_with_ordering_key.cc",
-    "internal/schema_auth.cc",
-    "internal/schema_logging.cc",
-    "internal/schema_metadata.cc",
+    "internal/schema_auth_decorator.cc",
+    "internal/schema_logging_decorator.cc",
+    "internal/schema_metadata_decorator.cc",
     "internal/schema_stub.cc",
     "internal/sequential_batch_sink.cc",
     "internal/session_shutdown_manager.cc",
     "internal/streaming_subscription_batch_source.cc",
-    "internal/subscriber_auth.cc",
-    "internal/subscriber_logging.cc",
-    "internal/subscriber_metadata.cc",
+    "internal/subscriber_auth_decorator.cc",
+    "internal/subscriber_logging_decorator.cc",
+    "internal/subscriber_metadata_decorator.cc",
     "internal/subscriber_round_robin.cc",
     "internal/subscriber_stub.cc",
     "internal/subscription_concurrency_control.cc",
