+++ google/cloud/pubsub/internal/publisher_logging_decorator.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "google/cloud/pubsub/internal/publisher_logging.h"
+#include "google/cloud/pubsub/internal/publisher_logging_decorator.h"
 #include "google/cloud/internal/log_wrapper.h"
 
 namespace google {
