+++ google/cloud/pubsub/internal/publisher_metadata_decorator.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "google/cloud/pubsub/internal/publisher_metadata.h"
+#include "google/cloud/pubsub/internal/publisher_metadata_decorator.h"
 #include "google/cloud/internal/api_client_header.h"
 
 namespace google {
